+++ src/main/java/com/yukgaejang/inflearnclone/domain/board/dao/UserDao.java
@@ -4,4 +4,5 @@ import com.yukgaejang.inflearnclone.domain.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserDao extends JpaRepository<User, Long> {
+    User findByEmail(String email);
 }
