+++ src/main/java/com/yukgaejang/inflearnclone/domain/board/dto/UserDto.java
@@ -8,5 +8,8 @@ import lombok.Setter;
 public class UserDto {
     private Long id;
     private String nickname;
+    private String role;
+    private String userName;
+    private String email;
 }
