--- src/main/java/com/yukgaejang/inflearnclone/domain/dao/UserDao.java
@@ -1,8 +1,0 @@
-package com.yukgaejang.inflearnclone.domain.dao;
-
-import com.yukgaejang.inflearnclone.domain.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserDao extends JpaRepository<User, Long> {
-    User findByEmail(String email);
-}
